requirement.txt
asgiref==3.2.10
Django==3.1.2
django-environ==0.4.5
pytz==2020.1
sqlparse==0.4.1
asgiref==3.2.10
cachetools==4.1.1
Django==3.1.2
djangorestframework==3.12.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
